Refactor modules to directories
